fix(db): gate syscall.Stat_t behind unix build tag
Windows builds have no syscall.Stat_t, so the cross-compile step in
GoReleaser broke on the v0.6.0 tag. Split describeDir into
describe_unix.go (//go:build !windows) and describe_windows.go, with
the Windows stub returning just path+mode (no POSIX uid/gid).

Unblocks release binaries on windows/{amd64,arm64}.
